WiP: Task and Project service

+++ src/main/java/com/bifee/projectmanagement/management/application/ProjectService.java
@@ -8,6 +8,8 @@ import com.bifee.projectmanagement.management.domain.project.ProjectStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.HashSet;
+
 
 @Service
 public class ProjectService {
@@ -18,11 +20,11 @@ public class ProjectService {
     }
 
     @Transactional
-    public Project createProject(CreateProjectRequest dto) {
+    public Project createProject(Long creatorId, CreateProjectRequest dto) {
         Project project = new Project.Builder()
                 .withTitle(dto.title())
                 .withDescription(dto.description())
-                .withOwnerId(dto.ownerId()).
+                .withOwnerId(creatorId).
                 withProjectStatus(ProjectStatus.IN_PROGRESS)
                 .build();
         return projectRepository.save(project);
@@ -44,9 +46,50 @@ public class ProjectService {
     }
 
     @Transactional
-    public Project updateProject(UpdateProjectRequest dto) {
-        Project project = getProjectById(dto.projectId());
+    public Project addMemberToProject(Long memberId, Long projectId, Long requesterId) {
+        Project project = getProjectById(projectId);
+        if(!isUserOwnerOfProject(projectId, requesterId)){
+            throw new IllegalArgumentException("Only owner can add members to project");
+        }
 
+        if (project.membersIds().contains(memberId)) {
+            throw new IllegalArgumentException("Member already in project");
+        }
+
+        var updatedMembers = new HashSet<>(project.membersIds());
+        updatedMembers.add(memberId);
+
+        Project updatedProject = project.mutate().withMembersIds(updatedMembers).build();
+        return projectRepository.save(updatedProject);
+    }
+
+    @Transactional
+    public Project removeMemberFromProject(Long memberId, Long projectId, Long requesterId) {
+        Project project = getProjectById(projectId);
+        if(!isUserOwnerOfProject(projectId, requesterId)){
+            throw new IllegalArgumentException("Only owner can remove members from project");
+        }
+        if (!project.membersIds().contains(memberId)) {
+            throw new IllegalArgumentException("Member not in project");
+        }
+        if (project.ownerId().equals(memberId)) {
+            throw new IllegalArgumentException("Cannot remove owner from project");
+        }
+
+        var updatedMembers = new HashSet<>(project.membersIds());
+        updatedMembers.remove(memberId);
+
+        Project updatedProject = project.mutate().withMembersIds(updatedMembers).build();
+        return projectRepository.save(updatedProject);
+
+    }
+
+    @Transactional
+    public Project updateProject(Long ProjectId, UpdateProjectRequest dto, Long requesterId) {
+        Project project = getProjectById(ProjectId);
+        if (!isUserOwnerOfProject(ProjectId, requesterId)) {
+            throw new IllegalArgumentException("Only owner can update project");
+        }
         Project.Builder builder = project.mutate();
         if(dto.title() != null){
             builder.withTitle(dto.title());
@@ -58,14 +101,27 @@ public class ProjectService {
             builder.withProjectStatus(dto.status());
         }
         return projectRepository.save(builder.build());
-
     }
 
-    // Fazer sistema de permissao para apenas o dono poder deletar um projeto
     @Transactional
-    public void deleteProjectById(Long id) {
+    public void deleteProjectById(Long id, Long requesterId) {
         Project project = getProjectById(id);
+        if(!isUserOwnerOfProject(id, requesterId)){
+            throw new IllegalArgumentException("Only owner can delete project");
+        }
         projectRepository.deleteById(id);
+    }
+
+    @Transactional(readOnly = true)
+    public boolean isUserMemberOfProject(Long projectId, Long userId) {
+        Project project = getProjectById(projectId);
+        return project.membersIds().contains(userId);
+    }
+
+    @Transactional(readOnly = true)
+    public boolean isUserOwnerOfProject(Long projectId, Long userId) {
+        Project project = getProjectById(projectId);
+        return project.ownerId().equals(userId);
     }
